use flint to replace fflas-ffpack
* It can be compiled on Windows now. using vcpkg to install FLINT and build it on Visual Studio.

* To use it, just copy the spasm.h and the compiled library.

* There're unknown bugs on schur random stuffs !!!!! (be careful) but the result is almost true.

* We provide a spasm_rref.cpp to connect rref stuff with Mathematica.

+++ src/spasm_triplet.c
@@ -151,7 +151,7 @@ struct spasm_csr *spasm_compress(const struct spasm_triplet * T)
 
 	/* success; free w and return C */
 	char mem[16];
-	int size = sizeof(int) * (n + nz) + sizeof(spasm_ZZp) * ((Cx != NULL) ? nz : 0);
+	ulong size = sizeof(int) * (n + nz) + sizeof(spasm_ZZp) * ((Cx != NULL) ? nz : 0);
 	spasm_human_format(size, mem);
 	fprintf(stderr, "%" PRId64 " actual NZ, Mem usage = %sbyte [%.2fs]\n", spasm_nnz(C), mem, spasm_wtime() - start);
 	return C;
